Make video recognition per mime type the standard
Videos where formerly recognized by a set of file extensions (mov, avi, mp4).

The new default search performs a mime type check. So if the file is of mime type `video/*`, then it is considered a video, independent of it's suffix.

Because this search type is not as performant as the search for suffixes, there's also the option to set the search type via -s.

Because the examples have reached an amount that would not fit into the usage message anymore, there's a new option -e to show usage examples. The usage message (-h) will only print one example.

Also converted the tabs to whitespace in the bash file.